Adds an administrative module to tweak the Log4Net threshold at runtime. You can already do this via your log4net.xml at runtime (it's set to FILE-WATCH), but only if you have immediate shell (or insert remote administration tool here) access.
One side effect is that you cannot at present use the module to set TRACE level - only DEBUG, INFO and WARN. This is due to the mechanism that we use for doing TRACE. 

Another side effect worth noting is that now the log file appends, where it did not before. This is due to the XmlConfigurator re-opening the log file after a rehash, which with append set to false causes the log file to truncate any activity to present, which is probably not advised. If you desire this, please override the log4net.xml file and tell the logger not to append. Lastly you might wish to change your startup scripts to do an "rm" or "del" on the spikelite.log so it doesn't grow out of hand. Or alter the rolling appender params, giving it a max size.

The console will be the appender that, by default, receives DEBUG messages. The log file still only receives INFO+ level, so if you're wondering where your lovely DEBUG level went then look at your console. Further, the new message interceptor for PRIVMSG works at the trace level - so you'll still have to modify the log4net.xml to get this.