Moved from Thresher to IrcDotNet.
Improved the console runner so it could handle a bot class that doesn't block.

Added the User class to replace a Thresher class that was being passed around inside the bot. The bot should now be totally decoupled from the irc lib types.

Fixed the post build event that copied sqlite to the output directory. It wasn't in quotes so failed on paths with spaces.

Changed the platform target to x86 for the FrontEnd_Console.

Fixed an issue where the bot was incorrectly reporting it's status during shutdown.