Refactored the IrcClient stuff to be properly decoupled from the rest of the bot.
Hooked up the IrcClient via the Spring.Net config.

Got rid of SpikeLite.Shared and merged all the classes in to SpikeLite.Core or SpikeLite.Domain.

Set both IrcCLient implimentation projects to output their assemblies to the main output directory.

Added a property ('Description') on IrcClient for displaying a description of the library it wraps.

Added missing licenses headers.

Lots of style and namespace clean up.